requirements.txt
streamlit==1.25.0
streamlit-drawable-canvas==0.9.3
pillow
pandas
reportlab
cairosvg
svgwrite
